Refactor api_construction to api
